APIS revamped
Revamped APIS to match the sample on Brightspace since we were pretty lost on how we were going to implement our own. I did make the more specific details more general like Katherine suggested with the use of other interfaces so we have more flexibility later if we want to change anything.

- also added a couple different folders so it is easier to look at our code. (tests and source). Im not sure if this is standard, so we can change it later